% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_AST}
\alias{btESBL_AST}
\title{Data: Antimicrobial sensitivity testing for a suvet of isolates}
\format{
A dataframe with 609 rows and 8 variables
\describe{
\item{supplier_name}{Unique sample ID}
\item{Organism}{E. coli or KpSC}
\item{amikacin}{AST result to amikacin}
\item{chloramphenicol}{AST result to chloramphenicol}
\item{ciprofloxacin}{AST result to ciprofloxacin}
\item{cotrimoxazole}{AST result to cotrimoxazole}
\item{gentamicin}{AST result to gentamicin}
\item{meropenem}{AST result to meropenem}
}
}
\usage{
btESBL_AST
}
\description{
Results from disc antimicrobial sensitivity testing (AST) according
to BSAC guidelines for six antimicrobials. Only a subset of isolates
underwent AST: 442 E. coli and 167 K. pnemoniae sequence complex (KpSC).
}
\keyword{datasets}
