% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_CARD_qrdr_mutations}
\alias{btESBL_CARD_qrdr_mutations}
\title{Data: CARD described QRDR mutations}
\format{
A tidy data frame with 1293 rows and 4 variables:
\describe{
\item{gene}{QDRR gene (gyrA, gyrB, parC, parE)}
\item{variant}{Identified variant}
\item{lane}{Unique sample-sequencing run ID}
\item{genus}{Genus of sample (E. coli or K. pneumioniae complex)}
}
}
\usage{
btESBL_CARD_qrdr_mutations
}
\description{
Quinolone resistnce determinant region mutations that have been
associated with resistance in E. coli, downloaded from the
Comprehensive Antibiotic Resistance Database https://card.mcmaster.ca/
on 6 September 2021
}
\keyword{datasets}
