% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_contigclusters}
\alias{btESBL_contigclusters}
\title{Contig cluster membership}
\format{
A tidy data frame with 714 rows and 10 variables:
\describe{
\item{id}{Contig identifier}
\item{clstr_size}{Number of contigs in cluster}
\item{length}{Contig length, bases}
\item{clstr_rep}{Is this the cd-hit cluster representative sample (1=yes,0=no)}
\item{clstr_iden}{Identity (\%) of sample to cluster representative sample}
\item{clstr_cov}{Coverage (\%) of sample to cluster representative sample}
\item{gene}{ESBL gene}
\item{clstr_name}{cd-hit cluster identfier: gene.n where n = 1,2..N and N is total no. clusters for given gene}
\item{lane}{Unique sample-sequencing run ID}
\item{species}{Species of sample}
}
}
\usage{
btESBL_contigclusters
}
\description{
Contig cluster assignment of all ESBL gene containing contigs, as
determined by cd-hit
}
\keyword{datasets}
