% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_contigclusters_msa_alignments}
\alias{btESBL_contigclusters_msa_alignments}
\title{Data: Contig cluster multiple sequence alignments as PopGenome GENOME objects}
\format{
A named list of ten data GENOME objects, where name is cluster id.
}
\usage{
btESBL_contigclusters_msa_alignments
}
\description{
Multiple sequence alignments of ten largest contig clusters, generated
using minimap2, mapping all cluster sequences to the longest sequence (i.e.
the cluster representative sequenc from cd hit). Sequence alignments have
been generated as .fasta files from the minimap2 .SAM files, then the
fastas loaded using the PopGenome::readData function.
}
\keyword{datasets}
