% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_contigclusters_msa_paf_files}
\alias{btESBL_contigclusters_msa_paf_files}
\title{Data: Contig cluster multiple sequence alignment .paf files}
\format{
A named list of ten data frames, where name is cluster id. Columns are:
\describe{
\item{qname}{Query sequence name}
\item{qlen}{Query sequence length, bases}
\item{qstart}{Query start coordinate (0-based)}
\item{qend}{Query end coordinate (0-based)}
\item{strand}{‘+’ if query/target on the same strand; ‘-’ if opposite}
\item{tname}{Target sequence name}
\item{tlen}{Target sequence length, bases}
\item{tstart}{Target start coordinate on orignal strand}
\item{tend}{Target end coordinate on orignal strand}
\item{nmatch}{Number of matching bases in the mapping}
\item{alen}{Number of bases, including gaps, in the mapping}
\item{mapq}{Mapping quality (0-255, with 255 if missing)}
}
}
\usage{
btESBL_contigclusters_msa_paf_files
}
\description{
Multiple sequence alignments of ten largest contig clusters, generated
using minimap2, mapping all cluster sequences to the longest sequence (i.e.
the cluster representative sequenc from cd hit).
}
\keyword{datasets}
