% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_coregene_tree_esco}
\alias{btESBL_coregene_tree_esco}
\title{E. coli core gene tree}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_coregene_tree_esco
}
\description{
Midpoint rooted maximum likelihood core gene phylogeny for E. coli
}
\keyword{datasets}
