% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_coregene_tree_kleb}
\alias{btESBL_coregene_tree_kleb}
\title{K. pneumoniae complex core gene tree}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_coregene_tree_kleb
}
\description{
Midpoint rooted maximum likelihood core gene phylogeny for K. pneumoniae
complex
}
\keyword{datasets}
