% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_global_popPUNK_clusters}
\alias{btESBL_ecoli_global_popPUNK_clusters}
\title{Data: popPUNK cluster assignment of study E. coli isolates using global database}
\format{
A tidy data frame with 473 rows and 2 variables:
\describe{
\item{lane}{Unique sample-sequencing run ID}
\item{Cluster}{popPUNK cluster assignment}
}
}
\usage{
btESBL_ecoli_global_popPUNK_clusters
}
\description{
PopPUNK cluster assignment of study E. coli using the popPUNK database
derived using a global E. coli collection as described in Horesh et al
2021; see manuscript for details
}
\keyword{datasets}
