% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_globalst131_tree}
\alias{btESBL_ecoli_globalst131_tree}
\title{Data: E. coli ST131 global phylogeny}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_ecoli_globalst131_tree
}
\description{
Midpoint rooted maximum likelihood phylogeny for E. coli
incorporating all ST131 study isolates and all ST131 E. coli
from a study of global ST131
(Mcnally et al 2019 https://doi.org/10.1128/mBio.00644-19)
reconstructed from a multiple sequence alignment from mapping to reference.
(see manuscript for details)
}
\keyword{datasets}
