% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_globalst167_tree}
\alias{btESBL_ecoli_globalst167_tree}
\title{Data: E. coli ST167 global phylogeny}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_ecoli_globalst167_tree
}
\description{
Midpoint rooted maximum likelihood core gene phylogeny for E. coli
incorporating all ST167 study isolates and 108 ST167 context genomes from
Patino-Navarrete et al 2020 (doi:10.1186/s13073-019-0699-6) obtained from
multiple sequence alignment from mapping to reference.
(see manuscript for details)
}
\keyword{datasets}
