% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_globalst410_tree}
\alias{btESBL_ecoli_globalst410_tree}
\title{Data: E. coli ST410 global phylogeny}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_ecoli_globalst410_tree
}
\description{
Midpoint rooted maximum likelihood phylogeny for E. coli
incorporating all ST410 study isolates 327 ST410 context genomes
from Feng et al 2019 (doi:10.1038/s42003-019-0569-1) using multiple
sequence alignment from mapping to reference.
(see manuscript for details)
}
\keyword{datasets}
