% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_globaltree}
\alias{btESBL_ecoli_globaltree}
\title{Data: E.coli global context tree}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_ecoli_globaltree
}
\description{
Midpoint rooted maximum likelihood core gene phylogeny for E. coli
incorporating study and global context isolates (see manuscript for
details)
}
\keyword{datasets}
