% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_horesh_metadata}
\alias{btESBL_ecoli_horesh_metadata}
\title{Data: metadata of 10,146 global E. coli genomes from Horesh et al}
\format{
A data frame with 10,146 rows and 22 variables:
}
\usage{
btESBL_ecoli_horesh_metadata
}
\description{
Metadata of a global collection of 10,146 E. coli genomes
Taken from Horesh et al 2021
See original Horesh publication for details and variable definition
https://doi.org/10.1099/mgen.0.000499
}
\keyword{datasets}
