% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_musicha_metadata}
\alias{btESBL_ecoli_musicha_metadata}
\title{Data: metadata of 97 E. coli genomes from Malawi}
\format{
A tidy data frame with 97 rows and 10 variables:
}
\usage{
btESBL_ecoli_musicha_metadata
}
\description{
Metadata of a collection of 97 E. coli genomes selected for temporal
diversity and diversity in antibiogram
Taken from Musicha et al 2017
For full details see original Musicha publication
https://doi.org/10.1093/jac/dkx058
}
\details{
The phylogroup and ST here were determined for this study and were not
provided in the original metadata; the popPUNK cluster assignemt uses the
Horesh database

#' \describe{
\item{lane}{Unique sample-sequencing run ID}
\item{ID}{Unmique sample ID}
\item{Year}{Year of sample collection}
\item{Sample_type}{Type of clincial sample from which isolate originates}
\item{Age_group}{Age of patient from which sample orignated}
\item{Organism}{Organism}
\item{ST}{Multlocus sequence type as determined by ARIBA}
\item{phylogroup}{Phylogroup following Clermon scheme with in silico PCR}
\item{Cluster}{PopPUNK cluster assignment using the Horesh E. coli database}
\item{Accession}{NCBI Accession number}

}
}
\keyword{datasets}
