% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_st131_amr}
\alias{btESBL_ecoli_st131_amr}
\title{Data: AMR genes identified in global ST131 isolates}
\format{
A tidy data frame with 5494 rows and 2 variables:
\describe{
\item{name}{Unique sample-sequencing run ID or accession}
\item{gene}{Identified AMR gene sequence}
}
}
\usage{
btESBL_ecoli_st131_amr
}
\description{
Identified AMR genes using ARIBA and the SRST database.
See manuscript for details
}
\keyword{datasets}
