% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_st131_metadata}
\alias{btESBL_ecoli_st131_metadata}
\title{Data: metadata of global E. coli ST131 isolates}
\format{
A dataframe with 918 rows and 4 variables
\describe{
\item{Accession_number}{Unique sample-sequencing run ID or accession}
\item{Year}{Year of sample collection}
\item{Country}{Country of sample collection}
\item{clade}{ST131 clade as defined in Mcnally et al}
}
}
\usage{
btESBL_ecoli_st131_metadata
}
\description{
Metadata of a collection of 918 E. coli ST131 genomes
including 855 from a study of global E. coli ST131
(Mcnally et at 2019 https://doi.org/10.1128/mBio.00644-19)
and the remainder sequenced as part of this study
}
\keyword{datasets}
