% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_st167_metadata}
\alias{btESBL_ecoli_st167_metadata}
\title{Data: metadata of global E. coli ST167 isolates}
\format{
A dataframe with 167 rows and 13 variables
\describe{
\item{accession}{Sample accession}
\item{Uberstrain}{Uberstrain as defined by Enterobase}
\item{Name}{Sample ID}
\item{platform}{Sequencing platform}
\item{library}{library as defined by Enterobase}
\item{insert_size}{insert size}
\item{Experiment}{Experiment accession number}
\item{Source Niche}{Source niche as defined by Enterobase e.g. human , animal}
\item{Source Details}{Source details as defined by Enterobase e.g. species}
\item{Country}{Country of sample collection}
\item{Collection Year}{Year of sample colllection}
\item{ST}{Multilocus sequence type}
}
}
\usage{
btESBL_ecoli_st167_metadata
}
\description{
Metadata of a collection of 108 E. coli ST167 genomes
Patino-Navarrete et al 2020 (doi:10.1186/s13073-019-0699-6)
}
\keyword{datasets}
