% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_st410_metadata}
\alias{btESBL_ecoli_st410_metadata}
\title{Data: metadata of global E. coli ST410 isolates}
\format{
A dataframe with 327 rows and 5 variables
\describe{
\item{accession}{Sample accession}
\item{Experiment}{Experiment accession number}
\item{host}{Host species from which isolate arose, if appropriate}
\item{source}{Sample source}
\item{Country}{Country of sample collection}
\item{Collection Year}{Year of sample colllection}
}
}
\usage{
btESBL_ecoli_st410_metadata
}
\description{
Metadata of a collection of 327 E. coli ST410 genomes
from Feng et al 2019 (doi:10.1038/s42003-019-0569-1)
}
\keyword{datasets}
