% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_st410_plasmids}
\alias{btESBL_ecoli_st410_plasmids}
\title{Data: Plasmid replicons identified in global ST410 isolates}
\format{
A tidy data frame with 3308 rows and 2 variables:
\describe{
\item{name}{Unique sample-sequencing run ID or accession}
\item{ref_seq}{Identified Plasmid replicon}
}
}
\usage{
btESBL_ecoli_st410_plasmids
}
\description{
Identified plasmid replcions using ARIBA and the PlasmidFinder database.
See manuscript for details
}
\keyword{datasets}
