% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_exposures}
\alias{btESBL_exposures}
\title{Exposures of included participants}
\format{
A data frame with 1838 rows and 26 variables:
\describe{
\item{pid}{Unique patient identifier}
\item{assess_type}{Day of observation (day 0=enrolment day)}
\item{died}{1=died on day of current observation}
\item{hosp}{1=hospitalised on day of current observation, 0 =not}
\item{amoxy}{Exposure to amoxicillin on day of observation}
\item{genta}{Exposure to gentamicin on day of observation}
\item{azithro}{Exposure to azithromycin on day of observation}
\item{tb}{Exposure to first-line TB treatment on day of observation}
\item{benzy}{Exposure to penicillin on day of observation}
\item{fluco}{Exposure to fluconazole on day of observation}
\item{ampho}{Exposure to amphotericin on day of observation}
\item{quin}{Exposure to quinine on day of observation}
\item{chlora}{Exposure to chloraquine on day of observation}
\item{coart}{Exposure to artemether/lumefantrine on day of observation}
\item{cipro}{Exposure to ciprofloxacin on day of observation}
\item{coamo}{Exposure to co-amoxiclav on day of observation}
\item{arte}{Exposure to artesunate on day of observation}
\item{cotri}{Exposure to co-trimoxazole on day of observation}
\item{clinda}{Exposure to clindamicin on day of observation}
\item{doxy}{Exposure to doxycycline on day of observation}
\item{erythro}{Exposure to erythomycin on day of observation}
\item{acicl}{Exposure to aciclovir on day of observation}
\item{fluclox}{Exposure to flucloxacillin on day of observation}
\item{metro}{Exposure to metronidazole on day of observation}
\item{strepto}{Exposure to streptomicin on day of observation}
}
}
\usage{
btESBL_exposures
}
\description{
These data are recorded in long format. Covariate values are included
in columns, and there is one row for each day where a covariate value
changes or when the participant has their final observation (i.e date
of censoring). So, if a participant has a covariate value that changes
from 1 to 0 on subsequent rows, the covariate value should be
interpreted as being 1 up until the day of change
}
\keyword{datasets}
