% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_kleb_globaltree}
\alias{btESBL_kleb_globaltree}
\title{Data: K. pneumoniae complex global context phylogeny}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_kleb_globaltree
}
\description{
Midpoint rooted maximum likelihood core gene phylogeny for K.
pneumoniae complex incorporating study and global context isolates
(see manuscript for details)
}
\keyword{datasets}
