% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_kleb_malawi_allisolate_core_gene_tree}
\alias{btESBL_kleb_malawi_allisolate_core_gene_tree}
\title{Data: K. pneumoniae complex core gene tree of Malawian carriage and invasive isolates}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_kleb_malawi_allisolate_core_gene_tree
}
\description{
Midpoint rooted maximum likelihood core gene phylogeny for K.
pneumoniae complex incorporating all Malawian isolates
(see manuscript for details)
}
\keyword{datasets}
