% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_model1posterior}
\alias{btESBL_model1posterior}
\title{Posterior of fitted carriage model (constant antimicrobial effect)}
\format{
An S4 stanfit object (package "rstan") including model log-likelihood
}
\usage{
btESBL_model1posterior
}
\description{
Posterior of fitted longitudinal model of ESBL colonisation incorporating
an effect of antimicrobial exposure that is stepwise-constant and
finsihes immediately following exposure
}
\details{
See manuscript for interpretation of parameters etc.
}
\keyword{datasets}
