% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_model2simulations}
\alias{btESBL_model2simulations}
\title{Data: Results of simulations from the model posterior}
\format{
A data frame with 600,000 rows and 18 variables:
\describe{
\item{sim_run}{Simulation run ID}
\item{draw}{Unique ID of draw from posterior (1,2...1000)}
\item{time}{Time t in days}
\item{hosp_days}{Number of days of hospitalisation}
\item{abx_days}{Number of days of antimicrobial exposure}
\item{pr_esbl_pos_t0esblneg}{Probability of ESBL colonisation at time t conditional on no colonisation at time 0}
\item{pr_esbl_pos_t0esblpos}{Probability of ESBL colonisation at time t conditional on colonisation at time 0}
\item{pr_esbl_pos}{Overall estimated prevalence of ESBL colonisation at time t assuming initial prevalence 0.5}
\item{prev_hosp}{Time of cessation of prev hospitalisation (999=none)}
}
}
\usage{
btESBL_model2simulations
}
\description{
These simulations set the antimicrobial and hospital exposure to
arbitrary values, set starting probability of colonisation to 0.5 then
draw parameter values from the posterior of the fitted model
(btESBL_model2posterior) to generate probability of being ESBL colonised
a time t later
}
\keyword{datasets}
