% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_model3posterior}
\alias{btESBL_model3posterior}
\title{Data: Posterior of fitted carriage model (seperate ceftriaxone effect)}
\format{
An S4 stanfit object (package "rstan") including model log-likelihood
}
\usage{
btESBL_model3posterior
}
\description{
Posterior of fitted longitudinal model of ESBL colonisation incorporating
an effect of antimicrobial exposure that exponentially decays following
cessation of exposure, with ceftriaxone effect seperate from
all otehr antimicrobials
}
\details{
See manuscript for interpretation of parameters etc.
}
\keyword{datasets}
