% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_modeldata}
\alias{btESBL_modeldata}
\title{Data: Covariate data to pass to Stan longitudinal carriage model in dataframe format}
\format{
The data from btESBL_stanmodeldata in dataframe format
}
\usage{
btESBL_modeldata
}
\description{
See vignette for details. Each row encodes two ESBL observations a time
t apart and the covariate values in this time period
}
\details{
These data are set up to fit a model where the effect of antimicrobial
exposure has a time-varying exponential decay effect, and hospitalisation
has a time-varying stepwise constant effect
}
\keyword{datasets}
