% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_NCBI_phenotypic_bl}
\alias{btESBL_NCBI_phenotypic_bl}
\title{Data: NCBI phenotypic beta-lactamase classifications}
\format{
A tidy data frame with 1293 rows and 4 variables:
\describe{
\item{allele_name}{Allele name}
\item{protein_accession_}{Protein accession}
\item{nucleotide_accession_}{nucleotide accession}
\item{gene_name}{Gene name}
\item{curated_gene_product_name}{Curated phenotype}
\item{class}{Inferred class of beta lactamase from phenotype}
}
}
\usage{
btESBL_NCBI_phenotypic_bl
}
\description{
Classification of beta lactamase genes downloaded from NCBI at
https://ftp.ncbi.nlm.nih.gov/pathogen/betalactamases/Allele.tab
on 17 September 2021
}
\keyword{datasets}
