% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_participants}
\alias{btESBL_participants}
\title{Characteristics of study participants.}
\format{
A data frame with 425 rows and 32 variables:
\describe{
\item{pid}{Unique patient identifier}
\item{arm}{Study arm; 1=sepsis, 2=inpatient, no abx, 3=community}
\item{enroll_date}{Date of enrolment in study}
\item{calc_age}{Age, years, at enrolment}
\item{ptsex}{Self reported gender}
\item{hivstatus}{HIV status}
\item{recieved_prehosp_ab}{Received antimicrobials (except CPT) last 4wk?}
\item{pmhxrechospital}{Hospital admission last 4wk?}
\item{tbstatus}{Ever been treated for TB?}
\item{tbongoing}{If tbstatus=Yes, currently being treated for TB?}
\item{hivonart}{If hivstatus=Reactive, current taking ART?}
\item{hivart}{If currently taking ART, which regimen?}
\item{art_time}{If currently taking ART, months on ART}
\item{hivcpt}{If hivstatus=Reactive, currently taking CPT?}
\item{tobaccoyn}{Currently smokes tobacco}
\item{alcoholyn}{Currently drinks alcohol}
\item{highestedu}{Highest education attained}
\item{job}{Current employment, if any}
\item{householdadultsno}{Number of adults in participant's household}
\item{householdchildno}{Number of children in participant's household}
\item{toilet}{Type of toilet in participant's household}
\item{watersource}{Usual household water source for drinking}
\item{watertreated}{Is drinking water in the household usually treated?}
\item{electricityyn}{Electricity available in household?}
\item{fuel}{Usual cooking fuel}
\item{keepanim}{Are animals kept in the household?}
\item{keep.poultry}{If animals are kept, are poultry kept?}
\item{keep.goats}{If animals are kept, are goats kept?}
\item{keep.dogs}{If animals are kept, are dogs kept?}
\item{keep.cattle}{If animals are kept, are cattle kept?}
\item{keep.sheep}{If animals are kept, are sheep kept?}
\item{keep.mules}{If animals are kept, are mules kept?}
}
}
\usage{
btESBL_participants
}
\description{
ART - antiretroviral therapy, CPT = co-trimoxazole preventative therapy
}
\keyword{datasets}
