% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_plasmidreplicons}
\alias{btESBL_plasmidreplicons}
\title{Data: Plasmid replicons identified in sequenced isolates}
\format{
A tidy data frame with 2893 rows and 3 variables:
\describe{
\item{lane}{Unique sample-sequencing run ID}
\item{ref_seq}{Identified plasmid replicon sequence}
\item{species}{Species of sample}
}
}
\usage{
btESBL_plasmidreplicons
}
\description{
Identified plasmid replicons using ARIBA and the PlamidFinder database.
See manuscript for details
}
\keyword{datasets}
