% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_sequence_sample_metadata}
\alias{btESBL_sequence_sample_metadata}
\title{Data: Accession numbers and metadata of isolates sequenced for this study}
\format{
A tidy data frame with 676 rows and 12 variables:
\describe{
\item{accession}{Sample NCBI accession number}
\item{lane}{Unique sample-sequencing run ID}
\item{supplier_name}{Unique sample id}
\item{pid}{Unique participant identifier}
\item{arm}{Arm of study 1=sepsis, 2=inpatient, no abx, 3=community}
\item{visit}{Study visit: 0 = baseline, 1 = d7, 2 = d28, 3 = d90, 4 = d180}
\item{data_date}{Date of sample collection}
\item{enroll_date}{Date of patient enrolment to study}
\item{assess_type}{Number of days post-enrolment that sample was collected}
\item{hosp_assoc}{Is sample community. hospital associated, or recent dc (see manuscript for details)}
\item{hospoutcomedate}{Date of discharge from hospital. NA = never admittted}
\item{Cluster}{PopPUNK cluster assignment. Prefix K = K. pneumo, E = E. coli}
\item{number_of_contigs}{Number of contigs in assembly}
\item{N50}{N50 of assembly}
\item{ecoli_phylogroup}{E. coli phylogroup from in silico PCR}
\item{ecoli_pathogroup}{E. coli pathogroup (see manuscript for details)}
\item{species}{Isolate species either from API (E. coli) or Kleborate}
\item{kleb_k_locus}{K. pnemo complex inferred K locus from Kleborate}
\item{kleb_k_locus_confidence}{Kleborate defined K locus confidence}
\item{kleb_o_locus}{K. pnemo complex inferred O locus from Kleborate}
\item{kleb_o_locus_confidence}{Kleborate defined O locus confidence}
\item{kleb_YbST}{Klebsiella YbST allele as per Kleborate}
\item{kleb_CbST}{Klebsiella CbST allele as per Kleborate}
\item{kleb_AbST}{Klebsiella AbST allele as per Kleborate}
\item{kleb_SmST}{Klebsiella SmST allele as per Kleborate}
\item{kleb_rmpA}{Klebsiella rmpA allele as per Kleborate}
\item{kleb_rmpA2}{Klebsiella rmpA2 allele as per Kleborate}
}
}
\usage{
btESBL_sequence_sample_metadata
}
\description{
Metadata of all sequenced isolates
}
\keyword{datasets}
