% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_snpdists_esco}
\alias{btESBL_snpdists_esco}
\title{Pairwise SNP-distance matrix for all E. coli genomes}
\format{
A 473 x 474 tibble.
}
\usage{
btESBL_snpdists_esco
}
\description{
First column is sample names - all other columns are sample names
}
\keyword{datasets}
