% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_snpdists_kleb}
\alias{btESBL_snpdists_kleb}
\title{Data: Pairwise SNP-distance matrix for all K. pneumoniae complex genomes}
\format{
A 350 x 351 tibble.
}
\usage{
btESBL_snpdists_kleb
}
\description{
First column is sample names - all other columns are sample names
}
\keyword{datasets}
