% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_stoolESBL}
\alias{btESBL_stoolESBL}
\title{Data: Results of stool testing for ESBL}
\format{
A data frame with 1416 rows and 9 variables:
\describe{
\item{pid}{Unique patient identifier}
\item{lab_id}{Unique speciment identfier}
\item{arm}{Study arm; 1 = sepsis, 2 = inpatient, no abx, 3 = community }
\item{visit}{Study visit: 0 = baseline, 1 = d7, 2 = d28, 3 = d90, 4 = d180}
\item{data_date}{Date of sample collection}
\item{sample_type}{Sample type (stool or rectal swab}
\item{ESBL}{Phenotypic ESBL presence (Positive) or absence (Negative)}
\item{enroll_date}{Date of patient enrolment to study}
\item{t}{Number of days post-enrolment that sample was collected}
}
}
\usage{
btESBL_stoolESBL
}
\description{
Results of testing stool/rectal swab samples for presence of any ESBL
producer
}
\keyword{datasets}
