% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_stoolorgs}
\alias{btESBL_stoolorgs}
\title{Results of stool testing for ESBL}
\format{
A data frame with 1032 rows and 3 variables:
\describe{
\item{lab_id}{Unique speciment identfier}
\item{organism}{Organism identification}
\item{ESBL}{Phenotypic presence of absence of ESBL in organism}
}
}
\usage{
btESBL_stoolorgs
}
\description{
Organisms identified from culture of stool on CHROMagar
(selective agar for cephalosporin resistance)
}
\keyword{datasets}
