% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_amrgenes}
\alias{btESBL_amrgenes}
\title{Data: AMR genes identified in sequenced isolates}
\format{
A tidy data frame with 10822 rows and 3 variables:
\describe{
\item{lane}{Unique sample-sequencing run ID}
\item{ref_seq}{Identified AMR gene sequence}
\item{genus}{Genus of sample (E. coli or K. pneumioniae complex)}
}
}
\usage{
btESBL_amrgenes
}
\description{
Identified AMR genes using ARIBA and the SRST database.
See manuscript for details
}
\keyword{datasets}
