% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_ecoli_globalst167_tree}
\alias{btESBL_ecoli_globalst167_tree}
\title{Data: E. coli ST167 global phylogeny}
\format{
A "phylo" object ("ape" package)
}
\usage{
btESBL_ecoli_globalst167_tree
}
\description{
Midpoint rooted maximum likelihood core gene phylogeny for E. coli
incorporating all ST167 study isolates and all ST167 E. coli
downloaded from Enterobase on 1st March 2021 using multiple
sequence alignment from mapping to reference.
(see manuscript for details)
}
\keyword{datasets}
