% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_model2simulations}
\alias{btESBL_model2simulations}
\title{Data: Results of simulations from the model posterior}
\format{
A data frame with 600,000 rows and 18 variables:
\describe{
\item{time}{Number of days post enrollment}
\item{1}{Probability of being ESBL non-colonised at current time}
\item{0}{Probability of being ESBL colonised at current time}
\item{draw}{Unique ID of draw from posterior (1,2...1000)}
\item{hosp_days}{Number of days of hospitalisation}
\item{abx_days}{Number of days of antimicrobial exposure}
\item{p0}{Probability at time 0 of being uncolonised}
\item{p1}{Probability at time 0 of being colonised}
\item{abx_start}{Time of antimicrobial exposure start}
\item{abx_stop}{Time of antimicrobial exposure stop}
\item{prev_abx}{Time of cessation of prev antimicrobial exposure (999=none)}
\item{hosp_start}{Time of hospitalisation start}
\item{hosp_stop}{Time of hospitalisation stop}
\item{prev_hosp}{Time of cessation of prev hospitalisation (999=none)}
}
}
\usage{
btESBL_model2simulations
}
\description{
These simulations set the antimicrobial and hospital exposure to
arbitrary values, set starting probability of colonisation to 0.5 then
draw parameter values from the posterior of the fitted model
(btESBL_model2posterior) to generate probability of being ESBL colonised
a time t later
}
\keyword{datasets}
