% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btESBL_stanmodeldata}
\alias{btESBL_stanmodeldata}
\title{Data: Covariate data to pass to Stan longitudinal carriage model}
\format{
A named list of seven items.
\describe{
\item{N}{Number of rows of data}
\item{t}{vector of length n, Time between ESBL observations for that row}
\item{n_covs}{Number of covariates to include in model}
\item{covs_type}{define types of covariates}
\item{start_state}{Starting state for row ESBL present absent}
\item{end_state}{End state for that time period}
\item{covs_mat}{Covariate values ofr each included covariate}
}
}
\usage{
btESBL_stanmodeldata
}
\description{
See vignette for details. Each row encodes two ESBL observations a time
t apart and the covariate values in this time period
}
\details{
These data are set up to fit a model where the effect of antimicrobial
exposure has a time-varying exponential decay effect, and hospitalisation
has a time-varying stepwise constant effect
}
\keyword{datasets}
